+++ src/SymMFEA/components/trainer/loss.py
@@ -3,7 +3,7 @@ import numpy as np
 import numba as nb
 from ...utils.functional import sigmoid
 import os 
-loss_jit = nb.njit([nb.types.Tuple((nb.float64[:], nb.float64))(nb.float64[:], nb.float64[:]),
+loss_jit = nb.njit([nb.types.Tuple((nb.float32[:], nb.float32))(nb.float32[:], nb.float32[:]),
                     ], cache= os.environ.get('DISABLE_NUMBA_CACHE') is None, nogil=True)
 
 class Loss:
